Enhance UserTopic relation for #81
This model allows one user to create a relation between a user (either self or other) and a topic.
The relationship label is called :action. A numeric score is possible.
by_user is not always needed. (user_id, by_user_id, action) has a unique index.

(1) A user can "follow" a topic. by_user is null as user is sufficient.
(2) user A can "rate" user B on a topic. A and B may be same or different.

This will be the beginning of our assessment/credentials module.