Item.extract_opengraph_data handles too many requests error
When extracting Open Graph data from a YouTube URL downloading the
content of the page can raise an exception. In particular, when
YouTube is throttling requests, they respond with a
`429 Too Many Requests` error.

In the default case, `Item.extract_opengraph_data` logs an error and
returns an empty hash when an exception is raised. This brings that
same behavior to the extract YouTube condition.